Changes in inventory,add item according to branch,item stock according to branch,transport routes,transport vehicle,transport assign vehicle,hostel rooms,hostel room type,hostel ,Student Certificate ,generate Certificate ,Student id card,generate id card,staff id card,genereate staff id card,front cms event ,front cms gallery,front cms news,front cms media manager,front cms pages

+++ application/views/admin/hostel/createhostel.php
@@ -1,8 +1,8 @@
 <style type="text/css">
-    @media print
-    {
-        .no-print, .no-print *
-        {
+    @media print {
+
+        .no-print,
+        .no-print * {
             display: none !important;
         }
     }
@@ -27,7 +27,7 @@
                             <h3 class="box-title"><?php echo $this->lang->line('add_hostel'); ?></h3>
                         </div><!-- /.box-header -->
                         <!-- form start -->
-                        <form id="form1" action="<?php echo site_url('admin/hostel/create') ?>"  id="employeeform" name="employeeform" method="post" accept-charset="utf-8">
+                        <form id="form1" action="<?php echo site_url('admin/hostel/create') ?>" id="employeeform" name="employeeform" method="post" accept-charset="utf-8">
                             <div class="box-body">
                                 <?php if ($this->session->flashdata('msg')) { ?>
                                     <?php echo $this->session->flashdata('msg') ?>
@@ -39,20 +39,30 @@
                                 ?>
                                 <?php echo $this->customlib->getCSRF(); ?>
                                 <div class="form-group">
+                                    <label for="exampleInputEmail1"><?php echo $this->lang->line('branch'); ?></label><small class="req"> *</small>
+                                    <select id="branch_id" name="branch_id" placeholder="" type="text" class="form-control">
+                                        <?php foreach ($branch as $key => $value) {  ?>
+                                            <option value="<?php echo $value["id"] ?>"><?php echo $value["branch_name"] ?></option>
+                                        <?php } ?>
+                                    </select>
+                                    <span class="text-danger"><?php echo form_error('branch'); ?></span>
+                                </div>
+
+                                <div class="form-group">
                                     <label for="exampleInputEmail1"><?php echo $this->lang->line('hostel_name'); ?></label><small class="req"> *</small>
-                                    <input autofocus="" id="amount" name="hostel_name" placeholder="" type="text" class="form-control"  value="<?php echo set_value('hostel_name'); ?>" />
+                                    <input autofocus="" id="amount" name="hostel_name" placeholder="" type="text" class="form-control" value="<?php echo set_value('hostel_name'); ?>" />
                                     <span class="text-danger"><?php echo form_error('hostel_name'); ?></span>
                                 </div>
                                 <div class="form-group">
                                     <label for="exampleInputEmail1"><?php echo $this->lang->line('type'); ?></label><small class="req"> *</small>
 
-                                    <select  id="type" name="type" class="form-control" >
+                                    <select id="type" name="type" class="form-control">
                                         <option value=""><?php echo $this->lang->line('select'); ?></option>
                                         <?php
                                         foreach ($ght as $type) {
-                                            ?>
+                                        ?>
                                             <option value="<?php echo $type; ?>" <?php if (set_value('type') == $type) echo "selected=selected"; ?>><?php echo $type; ?></option>
-                                            <?php
+                                        <?php
                                             $count++;
                                         }
                                         ?>
@@ -61,12 +71,12 @@
                                 </div>
                                 <div class="form-group">
                                     <label for="exampleInputEmail1"><?php echo $this->lang->line('address'); ?></label>
-                                    <input id="amount" name="address" placeholder="" type="text" class="form-control"  value="<?php echo set_value('address'); ?>" />
+                                    <input id="amount" name="address" placeholder="" type="text" class="form-control" value="<?php echo set_value('address'); ?>" />
                                     <span class="text-danger"><?php echo form_error('address'); ?></span>
                                 </div>
                                 <div class="form-group">
                                     <label for="exampleInputEmail1"><?php echo $this->lang->line('intake'); ?></label>
-                                    <input id="amount" name="intake" placeholder="" type="text" class="form-control"  value="<?php echo set_value('intake'); ?>" />
+                                    <input id="amount" name="intake" placeholder="" type="text" class="form-control" value="<?php echo set_value('intake'); ?>" />
                                     <span class="text-danger"><?php echo form_error('intake'); ?></span>
                                 </div>
                                 <div class="form-group">
@@ -81,16 +91,17 @@
                         </form>
                     </div>
 
-                </div><!--/.col (right) -->
+                </div>
+                <!--/.col (right) -->
                 <!-- left column -->
             <?php } ?>
             <div class="col-md-<?php
-            if ($this->rbac->hasPrivilege('hostel', 'can_add')) {
-                echo "8";
-            } else {
-                echo "12";
-            }
-            ?>">
+                                if ($this->rbac->hasPrivilege('hostel', 'can_add')) {
+                                    echo "8";
+                                } else {
+                                    echo "12";
+                                }
+                                ?>">
                 <!-- general form elements -->
                 <div class="box box-primary" id="holist">
                     <div class="box-header ptbnull">
@@ -107,7 +118,7 @@
                                 <thead>
                                     <tr>
                                         <th><?php echo $this->lang->line('hostel_name'); ?>
-                                        </th  >
+                                        </th>
                                         <th><?php echo $this->lang->line('type'); ?>
                                         </th>
                                         <th><?php echo $this->lang->line('address'); ?>
@@ -120,13 +131,13 @@
                                 </thead>
                                 <tbody>
                                     <?php if (empty($listhostel)) {
-                                        ?>
+                                    ?>
 
                                         <?php
                                     } else {
                                         $count = 1;
                                         foreach ($listhostel as $hostel) {
-                                            ?>
+                                        ?>
                                             <tr>
                                                 <td class="mailbox-name">
                                                     <a href="#" data-toggle="popover" class="detail_popover"><?php echo $hostel['hostel_name'] ?></a>
@@ -134,13 +145,13 @@
                                                     <div class="fee_detail_popover" style="display: none">
                                                         <?php
                                                         if ($hostel['description'] == "") {
-                                                            ?>
+                                                        ?>
                                                             <p class="text text-danger"><?php echo $this->lang->line('no_description'); ?></p>
-                                                            <?php
+                                                        <?php
                                                         } else {
-                                                            ?>
+                                                        ?>
                                                             <p class="text text-info"><?php echo $hostel['description']; ?></p>
-                                                            <?php
+                                                        <?php
                                                         }
                                                         ?>
                                                     </div>
@@ -150,17 +161,18 @@
                                                 <td class="mailbox-name"> <?php echo $hostel['intake'] ?></td>
                                                 <td class="mailbox-date pull-right no-print">
                                                     <?php if ($this->rbac->hasPrivilege('hostel', 'can_edit')) { ?>
-                                                        <a data-placement="left" href="<?php echo base_url(); ?>admin/hostel/edit/<?php echo $hostel['id'] ?>" class="btn btn-default btn-xs"  data-toggle="tooltip" title="<?php echo $this->lang->line('edit'); ?>">
+                                                        <a data-placement="left" href="<?php echo base_url(); ?>admin/hostel/edit/<?php echo $hostel['id'] ?>" class="btn btn-default btn-xs" data-toggle="tooltip" title="<?php echo $this->lang->line('edit'); ?>">
                                                             <i class="fa fa-pencil"></i>
                                                         </a>
-                                                    <?php } if ($this->rbac->hasPrivilege('hostel', 'can_delete')) { ?>
-                                                        <a data-placement="left" href="<?php echo base_url(); ?>admin/hostel/delete/<?php echo $hostel['id'] ?>"class="btn btn-default btn-xs"  data-toggle="tooltip" title="<?php echo $this->lang->line('delete'); ?>" onclick="return confirm('<?php echo $this->lang->line('delete_confirm') ?>');">
+                                                    <?php }
+                                                    if ($this->rbac->hasPrivilege('hostel', 'can_delete')) { ?>
+                                                        <a data-placement="left" href="<?php echo base_url(); ?>admin/hostel/delete/<?php echo $hostel['id'] ?>" class="btn btn-default btn-xs" data-toggle="tooltip" title="<?php echo $this->lang->line('delete'); ?>" onclick="return confirm('<?php echo $this->lang->line('delete_confirm') ?>');">
                                                             <i class="fa fa-remove"></i>
                                                         </a>
                                                     <?php } ?>
                                                 </td>
                                             </tr>
-                                            <?php
+                                    <?php
                                         }
                                         $count++;
                                     }
@@ -170,27 +182,33 @@
                         </div><!-- /.mail-box-messages -->
                     </div><!-- /.box-body -->
                 </div>
-            </div><!--/.col (left) -->
+            </div>
+            <!--/.col (left) -->
 
         </div>
         <div class="row">
             <div class="col-md-12">
-            </div><!--/.col (right) -->
-        </div>   <!-- /.row -->
+            </div>
+            <!--/.col (right) -->
+        </div> <!-- /.row -->
     </section><!-- /.content -->
 </div><!-- /.content-wrapper -->
 
 <script type="text/javascript">
     var base_url = '<?php echo base_url() ?>';
+
     function printDiv(elem) {
         Popup(jQuery(elem).html());
     }
-    function Popup(data)
-    {
+
+    function Popup(data) {
 
         var frame1 = $('<iframe />');
         frame1[0].name = "frame1";
-        frame1.css({"position": "absolute", "top": "-1000000px"});
+        frame1.css({
+            "position": "absolute",
+            "top": "-1000000px"
+        });
         $("body").append(frame1);
         var frameDoc = frame1[0].contentWindow ? frame1[0].contentWindow : frame1[0].contentDocument.document ? frame1[0].contentDocument.document : frame1[0].contentDocument;
         frameDoc.document.open();
@@ -216,7 +234,7 @@
         frameDoc.document.write('</body>');
         frameDoc.document.write('</html>');
         frameDoc.document.close();
-        setTimeout(function () {
+        setTimeout(function() {
             window.frames["frame1"].focus();
             window.frames["frame1"].print();
             frame1.remove();
@@ -227,13 +245,13 @@
     }
 </script>
 <script>
-    $(document).ready(function () {
+    $(document).ready(function() {
         $('.detail_popover').popover({
             placement: 'right',
             trigger: 'hover',
             container: 'body',
             html: true,
-            content: function () {
+            content: function() {
                 return $(this).closest('td').find('.fee_detail_popover').html();
             }
         });
